+++ scilab/modules/boolean/sci_gateway/cpp/sci_and.cpp
@@ -52,7 +52,7 @@ types::Function::ReturnValue sci_and(types::typed_list &in, int _iRetCount, type
 
     if (_iRetCount > 1)
     {
-        Scierror(999, _("%s: Wrong number of output arguments: %d to %d expected.\n"), "and", 1);
+        Scierror(999, _("%s: Wrong number of output arguments: %d expected.\n"), "and", 1);
         return types::Function::Error;
     }