+++ scilab/modules/elementary_functions/sci_gateway/cpp/sci_log1p.cpp
@@ -38,7 +38,7 @@ types::Function::ReturnValue sci_log1p(types::typed_list &in, int _iRetCount, ty
 
     if (in.size() != 1)
     {
-        Scierror(77, _("%s: Wrong number of input argument(s): %d to %d expected.\n"), "log1p", 1);
+        Scierror(77, _("%s: Wrong number of input argument(s): %d expected.\n"), "log1p", 1);
         return types::Function::Error;
     }